requirements.txt
h5py==2.7.0
Keras==2.0.3
numpy==1.12.1
olefile==0.44
Pillow==4.1.0
PyYAML==3.12
scipy==0.19.0
six==1.10.0
Theano==0.9.0
